requirements.txt
pyautogui~=0.9.54
pillow~=10.4.0
imageio
keyboard~=0.13.5
numpy~=2.1.0
opencv-python~=4.10.0.84